Move setting of package archices out of custom There's other things I want to do that will help bootstrap a new emacs install on a new machine and I'll find it easier to follow what I'm doing with this if I control the archives I'm working with.
The main reason I want to do this, right now, is I want to ensure that
`use-package` is available as early as possible, and installed if it
isn't avaialble, such that I can stop tracking my (m|d)elpa installs
in GitHub and have each emacs handle this. This will require, however,
that `use-package` is installed (it's not part of GNU emacs yet).

So... there's a bootstrap issue here and I think this will help solve
it.