Remove a couple of key combos for undo
Like with M-n and M-p to move windows, the use of M-BS and C-z for undo seem
to have been intended to be a bit Windows-a-like, or similar to other
editors I've used in the past. But, to this day, I still only ever use C-_.
So let's free up those other keys for something else.