Remove fill-column-indicator
It was clashing with company-mode:

https://twitter.com/davepdotorg/status/866642177706917888

and while trying to work out who the bug should be reported it, I stumbled
on:

https://github.com/alpaker/Fill-Column-Indicator/issues/72#issuecomment-260121526

which would suggest there's a bigger issue here and I might run into other
problems. Really nice mode/idea, implements a thing I really like in
SublimeText, but doesn't look like it's a good mix for my usual setup.