Don't even use-package magit if we're not on a Unix
I don't think use-package's :if does quite what I want. On machines that
don't have git I don't even want magit to be package-installed, let alone
actually used or configured. It seems :if isn't the way to do this, so let's
try not even invoking use-package on such a machine.