Move listing of personal bindings to f12 f12
For some reason I could never remember it was on f7, which is awkward 'cos
then I can't get to the list to see what it was on. f12 f12 seems easier to
remember given I've bound so many other things to f12 followed by some other
range of strokes.