Remove scss-mode (from melpa)
scss-mode on melpa doesn't seem to work so well either. Indentation works
for all files but I was finding that font-lock was failing in a lot of
cases. Turns out, I'd not noticed, these days GNU Emacs has a scss-mode
built right in.

So let's use that. Seems to work just fine.