:bug: Adding highlight intend to prog-mode-hook correctly
This should happen as an init, not a config. Oddly, I've only ever noticed
this not working on my work GNU/Linux box, it never seems to be an issue on
my iMac. The odd part here is that the GNU/Linux setup should be the correct
one so goodness knows why it worked on my iMac when it should really fail.