:zap: Tweak Emacs 26's line number display
I've had the previous code in my init file for ages, and I've not been
testing Emacs 26 for quite some time. Now that that 26 has been released,
and now that I'm running on it, change how I turn on line numbers such that
they're only shown in modes that derive from text-mode or prog-mode.

There might be others to come, but this will do for now.