:zap: Ensure ido- searching is case-insensitive
This didn't appear to be a problem in Emacs 25.x, it was case insensitive
out of the box. As of 26.1 (or some change in my environment due to that
update) it seems like it's case sensitive. What's even more interesting is
that, looking at the ido source, it sets its case folding to be whatever
case-fold-search is, and I always set that to t.

Whatever is going on, let's just simplify things by setting it by hand to be
sure.