:zap: Ensure auto-fill-mode is on for text modes
I'm not sure if this is something that's changed with Emacs 26.1, but
suddenly I'm finding that text-mode doesn't have auto-fill-mode on any more.
Until now I've always had auto-filling enabled in text-mode.

So, whatever the cause, ensure it's on for text-mode (and also don't bother
setting things for org-mode any more 'cos it derived from text-mode anyway)