:sparkles: Try and make projectile a little less greedy
For my major work projects I tend to have the "root" one level below the top
of the repository. The root, in this case, generally being the level at
which I pipenv the virtual environment for a Python/Django project. But I
keep finding that projectile will get a bit greedy and find both locations.
So, given a root path of:

/home/davep/projects/foo/src

It'll find both:

/home/davep/projects/foo/src

and

/home/davep/projects/foo

as projects. This change is an effort to discourage this.

I suspect it'll also mean I'll have to add top-level markers by hand.
Although, in most cases, I want it to find the .git directory, but I don't
want it to do that for projects that have a Pipfile or a .projectile file.

I'm not 100% sure that this change will pull that off, but I think it's a
start.