:zap: Trying out a possible fix for a forge issue
Getting https://github.com/magit/ghub/issues/81 with both GitHub and local
GitLab on the new Macbook Pro at work. I *think* I had this workaround in
place somewhere in my configuration, but perhaps it was in the "local"
config on my old machine.

So here I'm adding it as a global/all-machine fix until I get to the bottom
of what's going on.