:lipstick: Switch the powerline separator to utf-8
Until now I've been happy with the default 'arrow setting. However, on the
Macbook Pro, I'm seeing the effect mentioned in this issue:

https://github.com/milkypostman/powerline/issues/99

Thing is, this is with a setup that already has, and has had for a long
time, ns-use-srgb-colorspace set to nil. Using the utf-8 version seems to
make things look a lot more tidy.

I might need to see if this works fine everywhere else (especially GNU/Linux
and Windows), but I'm going with this for now.