:lipstick: Blink the cursor after I've been idle a good while
The idea here being that, if I'm in a graphical environment, and if I've
been looking away from the editor window for a while, I might have lost the
cursor. This should help make it more visible. Meanwhile it means it won't
be blinking while I'm actually busy typing.

Note that it's only in a graphical environment because `blink-cursor-mode'
only works in a graphical environment. In a terminal it's up to the
terminal.