:wrench: Turn off dimming when switching away from frame
This is a temporary change as there appears to be a small issue in dimmer.el
where, when switching away from the frame, all windows get dimmed (which is
fair -- I can see this making sense and I like the idea), but when switching
back, all windows that are configured to be ignored by dimmer don't get
un-dimmed.

See https://github.com/gonewest818/dimmer.el/issues/43 for the issue I
raised about this.