:lipstick: Change neotree theme
Something has changed in either neotree or all-the-icons, and now the
directories in the tree look a bit rubbish. Now showing directory folder
icons rather than the arrows I had before. I'm sure that makes sense to some
degree, but it looks unattractive to my eye and also the icon is right up
against the directory name, making it harder to read.

So, until I figure out why and what I can do about it, let's go back to the
classic theme.