:sparkles: Don't let quelpa manage melpa
I'm using quelpa as an exception, for those times where I either want a
package that isn't in melpa and the like, or where I want to use a
development branch of some package to test things.

So this helps to ensure that it isn't trying to manage melpa too.