:sparkles: Giving pretty-mode a go
I'm not sure if I like it yet, but it's fun to try new thing. Mostly I'm
going to notice this in Elisp and Python. I should be okay with the changes
there. It might get interesting when messing with Julia (if it supports
Julia, not sure yet).

See https://github.com/pretty-mode/pretty-mode