:sparkles: Experiment with having moving.el take over C-a/e
For a long time, I've used moving.el by having the commands bound to Home
and End, but I've always kept C-a and C-e as the Emacs defaults. I'm
thinking it might be time to experiment with taking them over too.

I can't see an obvious downside at the moment.