:hammer: Tweak things so I can use different Mac ports of Emacs
Until now, I've always used the Emacs from https://emacsformacosx.com/ but I
thought it was time to have an experiment with the Homebrew-friendly port
that is available via:

https://github.com/railwaycat/homebrew-emacsmacport

There are some very subtle changes in how things are done, mostly to do with
the names of variables that control the control of modifier keys, and also
the default font. These changes help me keep things the same as I'm used to
between the two different Mac ports of Emacs.