:fire: Remove the mapping of fn to control
Couple of reasons: first and most important is that this clashes with the
Homebrew/RailwayCat Emacs in that it causes Home/End/PgUp/PgDn to act as
C-arrow keys; second (and following on from the first) is that I very rarely
use Emacs on a non-full Mac keyboard these days, so I can do without the
mapping.

It's a shame in a way; it had benefits in the other port of Emacs for the
Mac, and so was handy on the rare occasions that I was using Emacs on the
Macbook Pro/Air on my lap, but if it kills the RailwayCat version (which I'm
trying out and finding some apparent benefit with) there's little point in
hanging on to it.

I can always set up some sort of command that I run on an ad-hoc basis on
the rare occasion I need to be working on a smaller keyboard.