:lipstick: Change how and where empty lines are indicated
Rather than turn on empty line indication everywhere, and them make
exceptions, leave it off (as is the default) and just turn it on for modes
that make sense. In most cases all the modes I'd want it on in are derived
from text-mode or prog-mode, so set it there. There's going to be some
awkward exceptions (restclient-mode, I'm looking at you), but I'll deal with
them when they become obvious or annoying.