:hammer: Revert back to M-g being goto-line
Originally I'd always had goto-line bound to M-g, given I never use
goto-char and so didn't need M-g to be a prefix of anything -- having to M-g
g (or M-g M-g) was just annoying. And then for a while I diverted off into
using a mode that did a live preview of where you were going while doing a
goto-line. Eventually I decided I didn't like it, but forgot to revert to my
original setup.

I decided to try and see if I could live with the more vanilla Emacs
approach to this but, mostly, it just clashed with my muscle memory. So
let's go back to being kind to my muscle memory...