:sparkles: Add the omitting of files to dired-mode-hook
Before Emacs 27.1, this was working just fine without the explicit addition
of a call to dired-omit-mode in dired-mode-hook. Now, with 27.1, this seems
to be necessary.

I could go down a rabbit hole of trying to work out what the difference is,
and why, but really I just want to be getting on with what I'm getting on
with. So... fix and forget.