:fire: Drop beacon
I've always liked it, but at the moment I'm finding it a little flaky.
Looking at the issues I seem to be running into a couple of things that were
reported some time ago, and it seems the author doesn't have the time to
address them at the moment.

Sadly I'm not really in a position to dive into the code and try and figure
it out so... given it was nice but not necessary, time to live without it.