:sparkles: Add placeholder setup of lsp
This came about from me trying to get this working with a couple of major
Python/Django projects. However, for various reasons (in part probably down
to pipenv, but also down to me still needing to read up on the docs for pyls
or something), it was becoming too frustrating to get going and I had to
quit while I was not too far behind.

Another day... I really would like to get lsp up and going but I also can't
and won't do it at the expense of how I work. It needs to fit in as an extra
useful tool, not as something that dictates how I write my code or else.