:sparkles: Attempt to get pyls going with lsp-mode
This still isn't quite working as I'd like, or at least in a way I can fully
follow, but it's also not crashing and burning either. I'm struggling to get
lsp-mode/pyls to use the correct linter and to display the lint results all
the time, but at least I'm also getting the basic benefits of using lsp-mode
too. So let's go with "good enough" for now and work out from there.