:card_index: Admit defeat and use pydocstyle for now
I can't seem to convince lsp-mode to use pylint and nothing else for all the
linting stuff, so go with pydocstyle as the in-editor nagging system. Given
that the way pydocstyle lets you configure it with a file (which is really
unhelpful -- every name and location seems to be designed to be unhelpful!)
let's just configure it right inside Emacs itself.

At least this way I have my preferences stored in a location where they will
get replicated and I can document them.