:lipstick: Have ivy minibuffer highlight span window width (again)
It just to work like this "out of the box", but stopped at some point in the
past (at least a year ago, if not more). It's been bugging me ever since.
Finally cracked and did a search for someone else having the issue and this
does seem to be noticed and this seems to be a quick fix.

https://github.com/abo-abo/swiper/issues/2383#issuecomment-675445899