:art: Remove the older Python lsp config
I'm still unsure about if I should fo with lsp-pylsp rather than
lsp-pyright, but right now the latter is working out fine, even if I seem to
be missing some of the more pedantic stuff (which was useful).

I'm going to run with this for a while and see how I go.