:fire: Remove setup-compile
At least for now. I'm getting an odd error with this ever since I updated to
Emacs 28.2, where it says during autoload that setup-compile-mode isn't
defined -- which is weird because there never was and there isn't such a
thing. A quick dive in has left me scratching my head and I don't know if
it's my code, use-package, Emacs, or something else.

But... I don't really rely on this package any more so I might as well not
lose time chasing down that rabbit hole. Not for now anyway. Another day...