:hammer: Experiment with switching to eglot for lsp stuff
It's similar but different and I feel it's the way to go but there's more
experimenting to do; hence I've not nuked my lsp config just yet. Yes, I
know, comments as version control and all that stuff; but it would be handy
to carry the LSP config around for a moment so I can look back at it.

So far though... so good.