:sparkles: Have M-RET toggle fullscreen frame
Now that I've tried swapping over to Emacs Plus, and I've configured the
frames to not have a title bar, there's no buttons to go fullscreen (not
that I use it as such much any more).

There is a command for doing it, so let's bind that to something I'm likely
to remember.